[refactor] 불필요 네이밍 리팩토링 및 type 필드 추가
